pull out order gossip handler
